+++ packages/components/src/panel/test/row.tsx
@@ -10,13 +10,15 @@ import PanelRow from '../row';
 
 describe( 'PanelRow', () => {
 	it( 'should render with the default class name', () => {
-		const { container } = render( <PanelRow /> );
+		const { container } = render( <PanelRow children={ null } /> );
 
 		expect( container ).toMatchSnapshot();
 	} );
 
 	it( 'should render with the custom class name', () => {
-		const { container } = render( <PanelRow className="custom" /> );
+		const { container } = render(
+			<PanelRow className="custom" children={ null } />
+		);
 
 		expect( container ).toMatchSnapshot();
 	} );